+++ src/parser/enrichers/trait/dragonborn/BreathWeapon2024.js
@@ -64,4 +64,19 @@ export default class BreathWeapon2024 extends DDBEnricherData {
     ];
   }
 
+  // get override() {
+  //   console.warn(this);
+  //   const uses = this._getUsesWithSpent({
+  //     type: "race",
+  //     name: this.data.name,
+  //   });
+  //   return {
+  //     data: {
+  //       system: {
+  //         uses,
+  //       },
+  //     },
+  //   };
+  // }
+
 }